# frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Keel CRM API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/deals/{id}/contacts/add/": {
            "post": {
                "operationId": "deals_contacts_add_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "deals"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deal"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/deals/{id}/contacts/remove/": {
            "post": {
                "operationId": "deals_contacts_remove_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "deals"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deal"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/deals/{id}/products/add/": {
            "post": {
                "operationId": "deals_products_add_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "deals"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deal"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/deals/{id}/stage/": {
            "post": {
                "operationId": "deals_stage_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "deals"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deal"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/deals/count/": {
            "get": {
                "operationId": "deals_count_retrieve",
                "tags": [
                    "deals"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deal"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ContactRef": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "email": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "email",
                    "id",
                    "name"
                ]
            },
...
            "Deal": {
                "type": "object",
                "description": "Base read serializer: owner/tag/custom-field output resolved from batch-loaded context.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "pipeline": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Ref"
                            }
                        ],
                        "readOnly": true
                    },
                    "stage": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StageRef"
                            }
                        ],
                        "readOnly": true
                    },
                    "company": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Ref"
                            }
                        ],
                        "readOnly": true
                    },
                    "primary_contact": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ContactRef"
                            }
                        ],
                        "readOnly": true
                    },
                    "amount": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,16}(?:\\.\\d{0,2})?$",
                        "readOnly": true
                    },
                    "currency": {
                        "type": "string",
                        "readOnly": true
                    },
                    "exchange_rate": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,10}(?:\\.\\d{0,8})?$",
                        "readOnly": true
                    },
                    "amount_base": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,16}(?:\\.\\d{0,2})?$",
                        "readOnly": true
                    },
                    "probability": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "expected_close_date": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true,
                        "nullable": true
                    },
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StatusCffEnum"
                            }
                        ],
                        "readOnly": true
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "lost_reason": {
                        "type": "string",
                        "readOnly": true
                    },
                    "stage_entered_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "description": {
                        "type": "string",
                        "readOnly": true
                    },
                    "owner": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/MembershipRef"
                            }
                        ],
                        "readOnly": true
                    },
                    "tags": {
                        "readOnly": true
                    },
                    "custom_data": {
                        "type": "object",
                        "additionalProperties": {},
                        "readOnly": true
                    },
                    "line_count": {
                        "type": "integer",
                        "readOnly": true,
                        "default": 0
                    },
                    "products_total": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,16}(?:\\.\\d{0,2})?$",
                        "readOnly": true,
                        "default": "0.00"
                    },
                    "version": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "archived_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "amount",
                    "amount_base",
                    "archived_at",
                    "closed_at",
                    "company",
                    "created_at",
                    "currency",
                    "custom_data",
                    "description",
                    "exchange_rate",
                    "expected_close_date",
                    "id",
                    "line_count",
                    "lost_reason",
                    "name",
                    "owner",
                    "pipeline",
                    "primary_contact",
                    "probability",
                    "products_total",
                    "stage",
                    "stage_entered_at",
                    "status",
                    "tags",
                    "updated_at",
                    "version"
                ]
            },
...
            "MembershipRef": {
                "type": "object",
                "description": "Compact owner/author reference: never exposes email or role to record readers.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "display_name": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "display_name",
                    "id"
                ]
            },
...
            "Ref": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
...
            "StageRef": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "kind": {
                        "type": "string",
                        "readOnly": true
                    },
                    "color_token": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "color_token",
                    "id",
                    "kind",
                    "name"
                ]
            },
...
            "StatusCffEnum": {
                "enum": [
                    "open",
                    "won",
                    "lost"
                ],
                "type": "string",
                "description": "* `open` - Open\n* `won` - Won\n* `lost` - Lost"
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "keel_session"
            }
        }
    }
}